Serb

+++ src/db/db_userdata.py
@@ -6,18 +6,33 @@ def db_user_get(db, user_id: str):
     if user_data.exists:
         return user_data.to_dict()
     else:
-        return "ERROR"
+        return "User not in database"
     
 
 def db_user_set(db, user_id: str, user_data):
-    return db.collection("UserData").document(user_id).set(user_data)
+    user = db.collection("UserData").document(user_id)
+    if not user.get().exists:
+        return "User not in database"
+    
+    user.set(user_data)
+    return "valid"
 
 
 def db_user_add(db, user_id: str, user_data):
-    db.collection("UserData").document(user_id).set(user_data)
-    return db_trackers.db_user_init(db, user_id)
+    user = db.collection("UserData").document(user_id)
+    if user.get().exists:
+        return "User with that id already exists"
+    
+    user.set(user_data)
+    db_trackers.db_user_init(db, user_id)
+    return "valid"
 
 
 def db_user_delete(db, user_id: str):
+    user = db.collection("UserData").document(user_id)
+    if not user.get().exists:
+        return "User not in database"
+    
     db.collection("UserData").document(user_id).delete()
-    return db_trackers.db_user_delete(db, user_id)
+    db_trackers.db_user_delete(db, user_id)
+    return "valid"
